requirements.txt
biopython>=1.80
numpy>=1.23.0
matplotlib>=3.5.0
scipy>=1.8.0
argparse>=1.4.0
csv>=0.15.0